Restructured src folder & added numbers game component
