+++ src/components/NumbersGame/NumbersGame.jsx
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+
+function NumbersGame() {
+    return <h1>yeeh, it's the numbers game!!!</h1>
+}
+
+export default NumbersGame;
